interact/legal-document-analysis/requirements.txt
streamlit
fastapi
uvicorn
PyPDF2
pypdfium2
pytesseract
pillow
langchain 
langchain-text-splitters
python-multipart
langchain-ollama
utils
dotenv
weaviate
spacy>=3.7.0
python-dateutil>=2.8.0
plotly>=5.17.0
networkx>=3.2.0
pydantic>=2.0.0
requests
langchain==0.0.354
langchain-core==0.1.23
langchain-ollama